feat: add a light classifier in the tl_classifer.py
A *.h5 file containing the MobileNet architecture with its weights was added. Also, a *npy file containing the classes-prediction relation.
In the tl_detector.py the call for the Subscribers was changed. Be aware that
that could modify the behavior of the previous commit.
Some code to draw a circle in an image (corresponding to the prediction) was added for debugging purpose.