feat: script to make the second fine tuning
I made some modifications on other scripts as well. I add the outputs of running each step, in short:
`bottleneck_fc_model.h5` is the resulting model of training the top layer with the rest of layers frozen with the Imagnet weights.
`capstone model.h5` is the resulting model of retraining the last seven layers with the rest of layers frozen with the Imagnet weights.
`capstone model_02.h5` is the resulting model of retraining the last seventeen layers with the rest of layers frozen with the Imagnet weights.